Merged eb/gcell-wip2 rev 10130:10152 into trunk.
This makes several gcell related changes.

{{{
The first two are INCOMPATIBLE CHANGES:

(1) The gcell portion of the code base was reorganized.  As part of that
reorganization, the paths to the include files changed.  They are now
installed under PREFIX/include/gcell instead of directly in PREFIX/include.

This means that all includes of the form:

  #include <gc_foo.h>

should be changed to

  #include <gcell/gc_foo.h>


(2a) If you are directly using gcell-embedspu-libtool or the
$(GCELL_EMBEDSPU_LIBTOOL) variable in your Makefiles, the order of the
two command line arguments was switched.  It's now 

  $(GCELL_EMBEDSPU_LIBTOOL) input_file output_file

or equivalently

  $(GCELL_EMBEDSPU_LIBTOOL) $< $@


gcell-embedspu-libtool allows you to convert an SPE executable
into something that libtool will allow you add to a host shared library.


(2b) The name of the symbol created by gcell-embedspu-libtool is now
suffixed with _spx (SPE executable) to reduce the probability of name
collision.  If you have lines like this:

  extern spe_program_handle_t gcell_all;

in your code, you may have to change them to:

  extern spe_program_handle_t gcell_all_spx;



The following changes are enhancements and shouldn't break any
existing code:


(3) We now install two new pkg-config files, gcell.pc and gcell_spu.pc.
These can be used to assist in building gcell code that lives outside
the GNU Radio repository.  The first gives the include and library
paths for the PPE host code, the second is the same info for the the
SPE code.

There is also a new .m4 macro, GR_GCELL, contained in
config/gr_gcell.m4, that uses PKG_CONFIG_MODULES to fish out the
relevant variables.  If you've got standalone code that uses gcell,
you'll probably want to copy this macro (along with our version of
pkg.m4) to your tree and use it.  It sets the following variables:

  GCELL_CFLAGS
  GCELL_CPPFLAGS
  GCELL_INCLUDEDIR
  GCELL_LIBS
  GCELL_SPU_CFLAGS
  GCELL_SPU_CPPFLAGS
  GCELL_SPU_INCLUDEDIR
  GCELL_SPU_LIBS
  GCELL_EMBEDSPU_LIBTOOL


(4) make -j now works in the gcell directory (fixes ticket:242).
}}}



git-svn-id: http://gnuradio.org/svn/gnuradio/trunk@10153 221aa14e-8319-0410-a670-987f0aec2ac5