Expanded the usrp props structure into something that can describe any kind of usrp: ethernet, udp, usb... and the various transport arguments.
The props structure is passed into to all the usrp make method to identify a usrp.
The find method takes a props structure as a hint and returns a vector of matches.
These capabilities will become more filled-in as the host code expands for other usrps.

Also, added an officiall struct for the mac addrs to handle the string parsing in one place.
Did the same for an ip address which will be needed for the udp usrps.