Add new block gr.additive_scrambler_bb()
This block performs scrambling by XORing the input sequence with
the output of an LFSR.  Repeating this operation restores the original
sequence.

(This differs from gr.scrambler_bb(), which convolves the input sequence
with the LFSR output.)

The additive scrambler allows an optional bit count after which the LFSR
is reset to its initial seed.  This allows use with, e.g., packetized
fixed length payloads.