Add missing buffer allocator hint to gr_ofdm_sampler.cc
This change was needed because gr_ofdm_sampler derives directly from
gr_block, but functions (from the runtime's point of view) as a
decimator.  It takes many items in to produce a single output
item.  Without this change, values of fft_size > ~4096 would result
in the scheduler complaining that it could never fullfil the block's
request for samples.

When deriving from gr_sync_interpolator or gr_sync_decimator, their
constructors set the hint automatically.