Change pmt message handling interface in gr_basic_block.h
Change the API such that the message handler is now implemented with a
callback, not an overridden virtual function.  The callback is now set
using gr_basic_block::set_msg_handler, which will accept pretty much
any kind of callable.

This change allows us to split the machinery for message handling out
from the block inheritance hierarchy, and provides a foundation that
can be used to build or experiment with arbitrary message dispatching
techniques.