grc: the check for pygtk can fail with no X session
However, the module will still import and we can perform the version check.
This change to the cmake python module checker will make sure that the
module was actually imported, even if the act of importing throws.
I believe that this change to the checker is acceptable for all other modules.