pfb: bug fix for ccf version of the PFB clock sync to handle the d_osps parameter better.
The number of output items is now set based on the d_osps and the check in the work function is not needed (actually, that check was wrong and should never have worked correctly, anyway).

This commit also changes the number of samples needed in the QA code for the constellation receiver for a faster test.