log: Restructure config file handling for gr-log.
The xml config file is installed into prefix/etc/gnuradio.

A new LOG section is added to gnuradio-core.conf that has an item log_file to point to the log file to be used.

If the log_file is not found, we can now set the file to "" to use the BasicConfigurator for very simple console logging. This is what will be used during 'make test'.

In gr_fir_sysconfig_x86.cc, all std::cerr calls are replaced with gr-log to easily turn them on/off.