uhd: fix pmt tuple ref namespace issue
The pmt_tuple_ref should be prefixed with pmt:: just like all of the other calls.
I am uncertain as to why this is compiling (we are not using namespace pmt),
unless there is a hidden using namespace somewhere in a global header.
But it will be good to get the code right as a general purpose example.