volk: fix popcnt in volk_rank_archs
The naive ranking system uses a popcnt to see which impl matches the most archs.
GCC gives us a builtin popcnt. MSVC also has __popcnt, but it seems to be
segfaulting, probably because the rank archs.c is built arch indepdent.

So, a manual popcnt has been implemented for all non gcc builin cases.
This is superior because 1) it doesnt segfault, 2) it covers all compilers.

Also note that the implementation does not have to be fast,
this is a one-time lookup process while the kernel table is loaded.