add the possibility to set x axes limits on the histosink in gr-wxgui
The axes limits is setted with 2 variables: range and offset.
Internally in the block gr_histo_sink_f, the values ​of maximum and
minimum of the histogram are checked in each frame.
These new two variables on the other hand will force the new desired
values of the maximum and minimum.
If the values of the limits of frame that were searched are higher
than desired, these are used