Fix bugs in atsc_field_sync_mux.
Since this block inserts field syncs into the data stream, clearly
it needs to be a gr_block rather than a gr_sync_block.  Also, the
d_already_output_field_sync variable needs to be initialized and
the d_current_index variable is unusued and thus can be deleted.

With these fixes in place, I've successfully transmitted an MPEG
transport stream with a USRP N210 and received it on my television.