volk/rotator: Fix the volk_32fc_s32fc_x2_rotator_32fc_generic kernel
abs() is an int function ...

Call the right one depending on the language now.

Note that I explicitely did the selection here rather than add a lv_abs()
function in volk_complex.h because there is no GCC extension for abs() that
would return the right type, so I'd have to stick to the cabs() function
which would force a double cast back and forth.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>