math: updated QA code to test nan and inf inputs to fast_atan2f.
When calculating with nan/inf values for x and/or y, make sure they all return and don't segfault. The results of gr_fast_atan2f in these cases may be different than atan2. But since we are concerned about speed with this block over accuracy (especially in these instances), we just want to make sure it doesn't crash.