grc: adding advanced tab feature to set a block's alias.
Does not allow for setting individual block aliases underneath a hier_block.

runtime: updates block registery to support updating block's alias that's then accessible through alias().

Adds an update_symbolic_name to the block_registry to delete the old key and add the new one. The block_registry only keeps the symbol_name (which never changes during the lifetime of the block) and the latest alias name that was set.