runtime: add stream flags to io_signature
This adds a new parameter to the io_signature class to allow specifying
flags for each input or output port.  The initial use case for this
capability is to allow a block to indicate that it needs to allocate its
own buffer memory for its outputs, or to indicate to the runtime that it
needs to let this block allocate upstream connected blocks' output buffers.

The flags are a 32-bit field with the following bits defined, which may be
or'd as needed:

MEM_BLOCK_ALLOC - Indicate to the runtime that this block needs to allocate
the memory for a given output buffer or the output buffer of an upstream
connected block. Absent other flags, the block's custom allocator must
provide a contiguous block of doubly mapped virtual memory that aligns on
a machine page boundary.

MEM_BLOCK_SINGLE - Indicate to the runtime that the custom allocator can
only define a singly mapped buffer.  This will cause the runtime to
restrict calls to work() to areas that do not wrap.

MEM_BLOCK_ODD - Indicate to the runtime that the custom allocator can only
allocate non-page aligned memory.