digital: Major fixes for filtering and tagging in correlate_and_sync_cc_impl
This commit fixes some major interrelated flaws in correlate_and_sync_cc_impl.
Since these fixes are effected by (ab)use of history(), delaying the primary
output, and setting the output multiple and *not* setting the alignment, they
couldn't easily be separated.

This commit fixes:
- maintaining consistent correlator operation across calls to work()
- not being able to tag samples back in time in a previous call to work()
- several array bounds checks/guards

The major side effect of these fixes is that the primary output is delayed
by the length of the user provided known symbols/preamble.  Thus the
optional second output (correlator matched filter output) have its peak
aligned to the start of the matched preamble as it is emitted on the primary
output.