digital: Add a tag marking delay input to correlate_and_sync_cc, fix phase_est
Add an input to allow the user to specify how many samples after the
corr_start tag, the phase_est, time_est, and corr_est tags appear.

Also fix the com,putation of the phase_est tag's value.  It only
really make sense (to me) to compute it at the same time the envelope of the
correlation peaks.

Updated the documentation and comments and added a reference to a paper
that provides some insights into complex cross-correlations.