blocks: message strobe shutdown fix.
On embedded systems, this block's destructure was not getting called
in the right order and the flowgraph would hang because this block's
thread running run would never quit. Moving the start and stop
handling of this thread to the start/stop functions of the block fix
this. Also makes sure that the block is only producing messages when
the flowgraph is running.