vocoder: possible fix for public header codec2.h
gnuradio/vocoder/codec2.h includes internal header in lib/ for defines.
This include avoids duplication of those defines for the in-tree build,
but it fails compilation when building against the installed headers.
This fix assumes that the ability to build against the installed header
outweighs the duplication cost of the #defines in the internal codec2.h