filter: fixes stability issue with IIR filters.
Addresses issue #773. IIR filters of some type combinations do not
appopriately handle the accumulator precision that can lead to
unstable results. This adds an accumulator data type template to
specify the precision that is needed internally here. The accumulator
should be of the highest precision data type used in the filter.