qtgui: fixes issue #767.
If FFT Size is set to 2048 and the half-width is selected for the
float graph type, then this would cause the x-axis display to show
-fs/2 to +fs/2 instead of the correct 0 to +fs/2. The problem was the
default numpoints was 1024, so with half 2048, the resetting of the
x-axis call was never hitting. Setting to 0 to force updates during
the first display.

Also puts protections around the allowable FFT sizes.