qtgui: enable/disable waterfall legend.
The waterfall plots have a different concept of the legend than the
others. It shows the intensity, but not the plot names. This allows us
to enable or disable showing the intesity as the right axis by setting
the Legend to Yes or No in the GRC properties box (or calling
disable_legend() on the waterfall object).