blocks: fixed problem with setting peak detector's alpha value.
This changes the public interface from an int to a float. But the
alpha must be between 0 and 1 for it to make any sense, so if this is
being used, it's never producing the right answer. Also, it's likely
that this change will be transparent to anyone who is trying to set
it.