gr-fec: trying to fix the volk submodule
On the previous commit, I accidentally mistook the seemingly empty volk directory for something that should be deleted, without realizing that it was the directory for the volk submodule. I've added the submodule back but I don't think it points to the same HEAD as the upstream configuration, and I can't figure out how to change that.

Hoping that it can just be ignored in the upcoming pull request.