Avoid enforcing unnecessary build constraints
Placing a build constraint in a CMakeLists.txt file like this can complicate build processes for downstream users.  While out-of-source builds are the modern way to build source code, in-source builds are still nevertheless a legitimate way to build code.  In this case, a warning message to the user is likely a more appropriate way to convey to a user the fact that they are using an outdated build strategy.