modtool: Fix bug where gr_modtool rm messes up the C++ QA section
This bug was originally reported by chenzhubattg. Previously, C++
QA files weren't properly removed from lib/CMakeLists.txt, such
that the path to the file was left in the CMake file, making it
invalid.
This fixes a bug in the cmake_editor.py module, which previously
had a too lenient regex.