Re-add qa_gr_flowgraph to test suite
During the great shuffling of v3.6->v3.7 qa_gr_flowgraph got dropped on the cmake floor.
This commit adds it back in (this lives in gr-blocks since it requires the nop/null_source/null_sink blocks).
This needed updating to the v3.7+ API (i.e. using gr::<blah> namespace instead of gr_<blah>), also two of the tests were failing since the null_source and null_sink have since been changed to
allow infinite outputs/inputs (i.e. the test that checks the flowgraph connect checker was failing due to the underlying block allowing multiple ports).
Instead we have to replicate the functionality of the null_sink/source in this file, but only allow a single in/out.
Also removes sine_table.h which lives in gnuradio-runtime since the great re-shuffling of 2013 (the only place it is used in in gnuradio-runtime/lib/math/fxpt.cc, and it does not get installed)