runtime: Make gr::buffer::allocate_buffer() public, not called by ctor
I've broken up the gr::buffer allocation to be separate from the constructor. This was already conveniently encapsulated in the allocate_buffer (private) virtual function.
This means that anyone creating a gr::buffer must now make a separate call to allocate it - fortunately this is relatively constrained within the scheduler to flat_flowgraph, but portaudio_source/sink also used a gr::buffer directly.