analog: Addresses issue #831.
We changed from using set_output_multiple(interpolate) because data
would not necessarily continue to flow through properly. This meant
that we needed to check the input size in the work function because if
we cycle through the filters, we can sometimes consume more inputs
than we've been given, which messes up the consume/produce and
bookkeeping.

It's likely that we can find a better form of the forecast function
here that does not require checking both noutput_items and
ninput_items[0] in the while loop.