digital: Addresses issue #812.
If the taps are all the same value, the differential taps become NaNs
during create_diff_taps. Having this type of filter doesn't make sense
in this algorithm, so there's nothing to fix. But now, we check to see
if the diff taps are NaN and throw a runtime_error if detected.