runtime: addresses issue 768.
Because of the data types in use here, d_ninput_items_required can
overflow and become negative, which then causes bad addressing of the
input buffer in a block's work function. Even if we were using
unsigned types, we might still have situations of overflow; it
wouldn't crash in this case but cause wrong addressing, anyways.

We might be able to catch and handle this situation better. This fix
tests for the condition and shuts down the flowgraph with an error
message. Basically, parameters that cause this effect are well out of
bounds of what we want to handle, anyways.