gnuradio-runtime/hier_block2: Allow changing of IO sig in the constructor
Fixes #719

The issue is that the hier_block2 detail creates some vectors for the
in/out ports to hold where to connect them to during the flatten and
what the min/max output buffer size are.

But subclasses are allowed to change the io signature in their constructor
so before actually using those vectors, we recheck if we should update
their size.

If there is a new io signature with more port, they're initialized to
default values (i.e. not connected). And if the new signature has less
ports, we just drop the connection (which is fine since it's all in the
subclass constructor, before they've been used for anything).

In both cases, just using .resize() on the vector is enough.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>