digital: addresses issue #890.
The set_taps was redone a while ago to better handle the differential
taps as well as the normal ones, but this made externally resetting
them impossible. Adding a new update_taps function to properly handle
this case so as to not change the API. We should fix this in 3.8.