blocks: fix incorrect use of items/bytes in PDU handling
For a uniform vector, calls to pmt::length return the number
of items in in the vector.  However, in some places this was
being treated as a byte value, resulting in short writes and
other misinterpretations.

These calls have been replaced with calls to pmt::blob_length,
to unambiguously get the length in bytes, then converted if
necessary to item values.