CMake: Fix dummy command in UseSWIG.cmake
Having the COMMAND parameter of the add_custom_command() function as an
empty string only works on Linux when generating Makefiles. It did not
work on Linux when using a different generator (in my case, Ninja), or
when building under Visual Studio.

This commit removes the empty string, and just don't specify any
command. This will work on all system / build system couples possible.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>