gr-fec: Address missing block error in example GRC file.
As-is, this flowgraph gives a "Missing Block" error. This is because it's looking for variable_ldpc_gen_mtrx_encoder_def, which we changed the name of during integration prior to pull request #660. I've updated this example GRC file to reflect the correct name of the intended block, variable_ldpc_encoder_G_def, which gets rid of the "Missing Block" error.

With this correction, the flowgraph can be executed, but it gives me the following error:

return _fec_swig.ldpc_encoder_make(*args, **kwargs)
TypeError: in method 'ldpc_encoder_make', argument 1 of type 'std::string'

I will address this separately as I don't think the error is caused by this GRC file.

Also, I've changed the block name of the other LDPC encoder definition block in this flowgraph to be the actual block that I intended to use.