digital: fixes how burst shaper produces full bursts.
In burst mode, the burst shaper wouldn't always produce a full burst
(through the end of the ramp down and post padding) because it was
requesting input in the forecast function for all states.

Changes forecast to not need input if in the end stages where its
inserting phasing (1/-1) samples and the post padding of 0's.

Work must check these states, too. If we don't use input, don't check
the nread condtition.