grc: Allow to use 'id' type for parameters
Currently the type 'id' can only be used for the actual 'id' field.
But it is useful to be able to use it for other things if you want
to reference a block from another one.

The only part where the current code doesn't work is when doing the
evaluation because it uses the type only to check for ID uniqueness.

This change uses both the type and the key. If the key is 'id' then it
must be unique from all other key 'id'. But if the key is something else,
then there must be an existing block to reference.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>