blocks: Add tag share block
The tag share block is used to transfer or share tags from one stream to another. The block takes two inputs and one output. The Input 0 stream is propagated to Output 0. In addition, the tags from Input 1 are added to the Input 0 stream. This is useful for transferring a detection tag from a correlate_access_code_bb block to the main complex stream, for example.