blocks: Add concatenate block
This block takes the input from several sources and put all the items
after each other. All the items from the source connected to the
first input are produced first, then all the items from the second
source, and so on. The upstream blocks need to return -1 in the
general_work method to make this block aware of when they have
completed producing items.

It can be useful for appending a vector (vector_source) to the end of a file
(file_source), appending one file to another file etc.