Fix additive scrambler tag triggered reset behavior
Update the additive scrambler to reset its internal LFSR on the
correct byte when operating in tag triggered mode. The LFSR should be
reset at/before the byte associated with the reset tag instead of
after it. The existing reset behavior is still correct for byte count
triggered reset.

Fixes #1198