runtime: fix logger_ptr to always be a pointer
Addresses issue #1383, where if GR and an OOT are built with logging
enabled but GR is build using log4cpp and the OOT is build without it,
then gr_block will be of different size & freeing it can result in
heap corruption.

This fix mades all of the logger_ptr typedefs an actual pointer, which
should always be the same size and hence gr_block will always be the
same size no matter what logging option is selected.