runtime: Use multiprecision rational arithmetic to propagate tags
Avoid floating point precision problems in propagation of tags for
the following cases:

- Common relative rates that are non-terminating in binary (1/3, 1/5, 1/7, ...)
- Large tag offsets
- Inexact floating point reciprocals when specifying the relative rate

This change to the runtime only fixes the tag propagation problems for
blocks with a fixed/unchanging relative rate.

This change adds a dependency on the MIPR library (a fork of the GMP
library) to perform multiple precision interger rational arithmetic.
MIPR was chosen as it purports to maintain a properly working Windows
build for cross-platform compatability.  Optimizations in MIPR are
also targeted to ARM in additional to x86-64.