cmake: fix CTRLPORT includes and testing
Currently, CTRLPORT_BACKENDS is not declared if ENABLE_GR_CTRLPORT is FALSE, so testing this variable makes no sense. All testing in gr-* is done using ENABLE_GR_CTRLPORT, so move to that in gr-blocks.

That said, having CTRLPORT_BACKENDS set to 0 could be useful, so rearrange CTRLPORT determination to set this variable correctly.

If ENABLE_CTRLPORT_THRIFT, then correctly use include_directories for header directories, -after- CTRLPORT is determined, not before. We can also combine some of these declarations to simplify the code.