Fixed wrong coefficient in Taylor series; cleanup
To even edit this file, indents had to be sanitized a bit. Also,
commented-out code was removed. Also, 8.f is a constand and not really
longer than `d_8`.

This is the first in two changes necessary for #1817. The other would be
adding a unit test that is *slightly* more holistic than just doing a
single-tone frequency detection.

Note that there's much about this code that could be refactored,
starting with Sdot being a FIR-filtered version of the input.