Remove catch statements from thread_body_wrapper.h
This commit removes the catch statements from thread_body_wrapper.h.
The reason to do this is that, as-is, a single block in a flowgraph
may throw an exception and stop functioning, but the only indication
of this to the user is printing the exception's what() to stderr.
With this patch, any uncaught exception will call std::terminate,
which should still print the exception's what(), but will also
terminate the process, rather than leaving a zombie flowgraph.