message: fix unit tests to respect mesage::to_string behaviour
Formerly to_string() returned a byte string in Python2 and a
unicode string in Python3.
To get the same behaviour for Python2 and Python3 the Unicode
String in Python3 is converted to bytes. Which is a desired behaviour
for to_string() since the carried payload of messages includes
invalid unicode characters.