interleave_impl.cc: Address Coverity 1327839, 1327840, 1327841, 1327842.
Dividing two ints gives a result that is an int. Adding 0.5f and converting
back to an int gives the original result. This change keeps the current
behavior and doesn't implement the (suspected) desired behavior. If
rounding is desired, the divison should be done as a float, so the 0.5f is
added to a float before truncation.

Signed-off-by: Philip Balister <philip@balister.org>