cmake: qwt: Only allow qwt-qt5 as a library name
The only way to be sure that FindQwt.cmake is finding the Qt5 version
of Qwt is to either
- limit the name to qwt-qt5 or
- have the user specify it when running CMake.

The latter is unimpeded by this change. However, by default, it'll now
skip libqwt.so on Fedora and Ubuntu systems, which is actually only
a valid library when using Qt4, not Qt5.

Fixes #2027.