trellis interleaver: made table size unsigned int
This is part of the ongoing effort to avoid signedness confusion.

It completes df4e7a85263eb523395ce6bef0b2e7769bd63465 in that it at
least strives to handle sizes consistently in interleaver.cc.

Of course, the returned vectors still contain signed ints where there
should be unsigned ints, but the trellis core_algorithms.cc code is
convoluted enough to warrant carefulness when touching, so this part of
the API was kept.