gr-vocoder: actually use LIBGSM_INCLUDE_DIR if a gsm.h was found.
FindGSM.cmake looks for the header and sets the LIBGSM_INCLUDE_DIR,
which was not used.
The build worked anyhow, because gsm.h normally resides in a
system include directory.
Tested on Ubuntu 18.10 (/usr/include/gsm.h is a symlink to
/usr/includegsm/gsm.h) and NetBSD/pkgsrc (/usr/pkg/include/gsm.h
is the real one, not in the gsm/ subfolder).