Reduce peak memory usage of Octave function read_cshort_binary
Attempts to read large amounts of data with 'read_cshort_binary' may
make Octave 6.0.0 overflow and exit with the message:
'octave exited with signal 6'.
The purpose of this change is to use more sparingly memory in order
to avoid overflowing in cases where the computer *does* have enough
memory to perform the desired operation.