grc sometimes generates wrong yml description of  hier2 block
If  a hier2 block contains more than one ( input or output ) message port
grc generates a wrong yml description of this block.

If you restart grc , grc terminates with an error messages is no longer usable.

You can check with the example in

gr-blocks/examples/msg_passing/hier

This example is running in 3.7 without any problems.

The modification in platform.py avoids the termination of grc
if there is an error in the flowgraph to be loaded.

The modification fixes the bug in the generated yml code.