cmake: Enable CMAKE_EXPORT_COMPILE_COMMANDS
When CMAKE_EXPORT_COMPILE_COMMANDS is ON, a file compile_commands.json
will be created in the build directory that enables code completion
engines to find all include directories and provide a very comprehensive
auto completion experience without any manual IDE configuration.

A nice example for this is the VS Code extension cpptools that is
developed by Microsoft. Of course this is not limited to VS Code and
works with other engines such as YouCompleteMe as well.