uhd: Create compatibility layer for future UHD versions
clock_config_t and io_type_t are UHD types that have been flagged as
deprecated, in some cases since 2012. UHD 4.0 will finally remove those
types.

In order to stay compatible with future versions of UHD without breaking
the GNU Radio APIs on this API-stable branch, we add the types directly
into GNU Radio. The implementation is such that the UHD version will be
used if available, and the GNU Radio version will be used otherwise.