digital: examples: Correct framebits uhd_packet_rx*.grc
The example grc files have an incorrect parameter for framebits in the
dec_hdr block. This is correct in packet_loopback_hier.grc. The files
uhd_packet_rx.grc and uhd_packet_rx_tun.grc both contained this error
since creation. This error causes the blocks to get in sort of
a deadlock when receiving the header with the dummy FEC.