digital: Fix GRC examples for YAML format
With this commit the remaining GRC examples that were not yet converted
to YAML are converted too.

GRC did not open these examples because it threw an error.

> Error: (ValueError("invalid literal for int() with base 10:
> 'digital.packet_utils.default_access_code'",), '% if
> int(eval(access_code))==0:\nself.${id} = ${id} =
> digital.header_format_default(digital.packet_utils.default_access_code,\\\n${threshold},
> ${bps})\n% else:\nself.${id} = ${id} =
> digital.header_format_default(${access_code},\\\n${threshold}, ${bps})\n% endif')

This could be fixed by manually editing the corresponding GRC files. All
uses of `digital.packet_utils.default_access_code` are replaced with a
string. This allows GRC to open these files. Still, GRC will hang in
case `digital.packet_utils.default_access_code` is used again.
This hints at a bug with the `default_access_code` code.