Add 'Watch End Tag' option to 'ZeroMQ Pull Source' block
if 'Pass Tags' option of the 'ZeroMQ Pull Source' GRC block is set, a 'Watch End Tag' option will appear, that if enabled, watches incomming data for an item that is tagged with "end" key. The block will return WORK_DONE in its work function if all items before (and including) that tagged item is written to output port.