Don't use hard-coded path constants.
This replaces the path constants for the prefix, system configuration,
and preferences directories with a runtime prefix lookup based on the
gnuradio-runtime library location and relative paths from that.

Also use cmake-style paths instead of native paths for substitution to
avoid unescaped character problems. For relocatable prefix string
replacement, conda always substitutes in forward-slash paths anyway, so
it's just easier to use them from the start.