grc: add python snippets to GRC
This feature adds the ability to insert arbitrary code into the python
flowgraph.  It gives a little more low-level flexibility for quickly
modifying flowgraphs and adding custom bits of code rather than having
to go and edit the generated py file

One example is synchronizing multiple USRP objects - sometimes you want
different sync than what is offered in the multi-usrp object, so you can
put a bit of code in the snippet block to do the custom synchronization

Singleton block to contain all the snippets

Split _multiline and _multiline_external dtypes from string in the
processing as code block and snippet are getting evaluated multiple
times.  If the code contains sleep or print, this wreaks havoc.