digital: add filter response truncation to generic mod
The generic mod implementation is a convenience hier block to modulate
bits to symbols and apply an RRC filter.  One downside is the output is
delayed by the length of the RRC filter (which is specified inside the
block).  This adds an option to truncate the output according to the
length of the filter response such that the start of output is aligned to the first symbol.

Fixes #2920