grc: loosen the validation of block in function probe
The block_id parameter for the block being probed was being validated as
a block_id.  But there is a use case where this can be blank to call a
function on the flowgraph object itself.

Changed the block_id dtype to "name" just like function_name.  Could
have just changed it to raw and done no validation, but that doesn't get
displayed very nicely in the grc block

Fixes #2966