additive_scrambler: Add proper test for count based reset
Do not chain two additive scramblers to check whether the reset is
performed correctly: A flaw in the reset logic can not be detected this
way as the same logic is used both in the scrambler and the descrambler.

Instead, check whether the first N elements generated by the scrambler
are repeated in the output (where N is the reset count). This test does
fail as of this commit; see [1].

[1] https://github.com/gnuradio/gnuradio/issues/2926