additive_scrambler: Fix count based reset
Count based reset in the additive scrambler may be delayed by a single
byte as reported in [1]. This happens whenever the work() function of
the block is called: The first value of reset_index in work() is d_count
- d_bytes where d_count is the number of bytes to output before a reset
and d_bytes is the number of bytes produced after a previous reset (in
the previous work() call). The very first reset is delayed by one byte
due to zero-based counting in work()’s loop /and/ a check of the reset
condition (i == reset_index) /after/ outputting a byte. (All following
resets are performed at the correct interval after 4 bytes.)

Example: With d_count = 4 and d_bytes = 0 the first reset is performed
at i = 4, but only after a fifth byte has been output.

Fix this error by checking the reset condition /before/ outputting a
byte. This removes different handling of tag and count based reset,
introduced by commit e3ad82e6d9 as a fix for delayed tag based reset
[2]. The fix for tag based reset should have been applied to the count
based reset as well (resulting in the same code as of this commit). The
error in count based reset has not been detected due to wrong QA code
(see previous commit).

Fixes #2926.

[1] https://github.com/gnuradio/gnuradio/issues/2926
[2] https://github.com/gnuradio/gnuradio/issues/1198