Rotator-based Freq Shift Convenience Wrapper
This yml adds a convenience-wrapper block to provide volk-accelerated
frequency shifting based on the rotator block that doesn't require
the user to "import math" and code the "2Pifreq/samp_rate" as the
phase increment. I don't think newbies know a rotator with a phase
increment results in a frequency shift (I didn't for the longest time)
and may be relying on less CPU-efficient means such as a signal
source/multiply or a frequency xlating fir filter (the latter
one-sample-at-a-time rotates as it decimates rather than leveraging
volk to do a block).