audio: pre-allocate and introduce checks for alsa sink output buffer
In gr::audio::alsa_sink, allocate memory for output buffers in
check_topology(). Typecast input items in bulk using volk's
casting functions. Always check if the number of output
items has changed in work(), just to be safe. This change only
covers work_s32() but the rest will be trivial if this
is the desired method of operation.