Improve Performance of PFB Arbitrary Resampler
The code updates take advantage of the fact that the filter and
differential filter in the work loop can be computed independent
of each other.  A second thread is used to process the filter while
the differential filter is being calculated to cut the processing
time in half.  Valgrind measurements with and without the change
showed showed a call cost reduction from 41% to 14%.  A significant
increase in throughput.