cmake: Increase optimization level to 3 for gcc and clang
Note: RelWithDebInfo still uses -O2.

For gcc, increasing the optimization level adds the following
optimizations over the existing ones:

-fgcse-after-reload
-finline-functions
-fipa-cp-clone
-floop-interchange
-floop-unroll-and-jam
-fpeel-loops
-fpredictive-commoning
-fsplit-paths
-ftree-loop-distribute-patterns
-ftree-loop-distribution
-ftree-loop-vectorize
-ftree-partial-pre
-ftree-slp-vectorize
-funswitch-loops
-fvect-cost-model
-fversion-loops-for-strides

In particular, this includes loop optimizations and tree vectorization.
Some of these options can increase code size and most likely will
increase compilation times, but performance is preferable and code size
can be minimized by switching to MinSizeRel as the build target type.

In clang, the difference to -O2 isn't quite as big, but we keep the same
optimization level for symmetry.

This commit forces the default debug level to O3, independent of what
the local CMake setting is. For CMake versions that default to -O3, this
commit changes nothing.