cmake: Change debug optimization level to -Og for gcc/clang
gcc and clang used to force the default optimization level for debug
builds to -O2. This is unusual, and can make gdb-based debugging
complicated. This changes the default level to -Og, which is not the
CMake default, but is the recommended level for debug/compile cycles.

Note that RelWithDebInfo uses -O2, so there are now different targets
for different debug levels.