gr-blocks Add Msg to Var and Var to Msg Conversion Blocks
These 3 new blocks provide the missing glue to move between messages
and variables in the same flowgraph.  There are 3 variants here:
1. Monitor a variable and produce a user-specified message (pair)
when the variable changes.  Useful bridging standard GUI controls
to message-based blocks.
2. When an inbound message (pair) is received, update a specified
variable.
3. When an inbound message (dict) is received, extract a single
dictionary entry and produce a message pair (useful if you have a
multi-value dictionary but you just want to pull off a single
attribute such as frequency or gain without modifying the upstream
block. This can be paired with (2) if necessary to move from a
dictionary item to a variable.