file_sink: Flush file on stop()
Make sure that data passed to the file sink is written to the disk when
the flowgraph is stopped. Before this change this only happened
implicitly due to a fclose() in the base class destructor. This in turn
means that data was written delayed after stopping the flowgraph [1].

This fixes #2590.

[1] https://github.com/gnuradio/gnuradio/issues/2590