audio: replace stderr logging by calls to GR's logging facilties
This is a bit special, in that audio systems tend to have their own ways
of going wrong, and there's platform-specific development debugging
infrastructure in there, which I intentionally did not touch.

I did touch a few commented lines of code in the intention of, if you
enable that line of logging, it should be consistent with the rest.

There's copious amounts of untouched stderr-logging in `#if 0`-disabled
code in OSX's infrastructure.