utils: Make modtool generate more C++ style compliant code
This doesn't entirely make it perfect, since some formatting depends
on the length of the name of the block, and the length of its
parameters.

In the general case it can't be made perfect without reimplementing
`clang-format`, but even then the generated code is not actually valid
C++ (the size of the data is not defined at `gr_modtool add` time, so
uncompilable placeholders are put in place).

So this may be as close as we'll get generating the code.