cmake: use TIMESTAMP function for build date and generate constants once
Previously a a custom python command was used to generate the build date
but this can be done with native CMake functions for all supported CMake
versions.

Additionally it's compatible with reproducible builds since CMake 3.8 as
it honors SOURCE_DATE_EPOCH set in the environment varibales.

Thanks to Sebastian Kosloswki for pointing out the right way in CMake