blocks: Fix XMLRPC Client and Server block templates
The XMLRPC Client block is currently broken because it tries to import 
the Python 2 XMLRPC module, and because the "addr" and "callback" 
parameters are inserted as quoted strings. To fix these problems, I've 
switched to the Python 3 version of the XMLRPC module with a fallback to 
the Python 2 version, and changed the data types of the problematic 
parameters to "raw".

The XMLRPC Server block is currently broken because it tries to import 
the Python 2 XMLRPC module. To fix this problem, I've switched to the 
Python 3 version of the XMLRPC module with a fallback to the Python 2 
version.