uhd: Add PC clock resync command
USRPs may be synchronized to a host clock. Over time USRP and host clock
deviate. This is an issue for timed commands. With this commit it is
possible to resync both devices.

The new command key is `pc_clock_resync`. The command does not expect
any argument.

It may be possible that USRP time is not strictly monotonic anymore. If
that is a requirement, other options should be explored.

e.g. connect a 'Message Strobe' block to a USRP command port and send a
message with `pmt.dict_add(pmt.make_dict(), pmt.intern('pc_clock_resync'), pmt.PMT_T)`
periodically. It is sufficient to send such a message every couple of
seconds.