uhd: Fix missing bindings
uhd python bindings were missing a couple of functions.

1. `uhd.ALL_MBOARDS` was not defined.
2. `uhd.streams_args` was incomplete.
Instead of import magic, we define a pybind constructor which supports
all cases. This should ease the maintainance burden as well.

Be aware: some notes that `this file was generated with bindtool` were
removed because this files contain manual edits. The foundations were
probably generated with bindtool once.